To grub 2 yay

+++ kernel/multiboot/mb_info_block.h
@@ -24,7 +24,7 @@ struct mb_info_block {
     uint32_t vbe_interface_seg;
     uint32_t vbe_interface_off;
     uint32_t vbe_interface_len;
-    uint32_t framebuffer_addr;
+    uint64_t framebuffer_addr;
     uint32_t framebuffer_pitch;
     uint32_t framebuffer_width;
     uint32_t framebuffer_height;
